#793 Moved sequential edit action to StandardActions.
