+++ platform-web-resources/src/main/java/ua/com/fielden/platform/web/test/server/WebUiConfig.java
@@ -1285,7 +1285,7 @@ public class WebUiConfig extends AbstractWebUiConfig {
                                 shortcut("alt+n").
                                 build())
                 .also()
-                .addTopAction(CentreConfigActions.SEQUENTIAL_EDIT_ACTION.mkAction(TgPersistentEntityWithProperties.class))
+                .addTopAction(StandardActions.SEQUENTIAL_EDIT_ACTION.mkAction(TgPersistentEntityWithProperties.class))
                 .also()
                 .addTopAction(action(EntityDeleteAction.class).
                         withContext(context().withSelectedEntities().build()).
